Wrote failing test for tag+after subquery behaviour
The current solution for left-justified after queries (selecting from a subquery) causes problems because the subquery must be as comprehensive as the outer query. At the moment selecting 2 documents tagged with 'web' after a datetime, if there is a single note NOT tagged with 'web' after datetime before a document tagged with 'web', incorrect results are given.

This leads me to conclude that handling this behaviour within librarian is a futile attempt, and must be handled at app-level. Any further attempts will only complicate things further, and removing this behaviour will make the API much more explicit, with fewer confusing side effects to worry about.